First commit of an example adaptation.
Here is a short and non-extensive summary of the modifications;

 + pom.xml
   - changed source and target for Java 8

 + README.md
   - renamed original and created a new one

 + endit-tapebackend-interactions.png
   - a new file

 + PollingTask.java
   - added a new 'start(path)' method for FlushTask

 + StageTask.java
   - no sleeping

 + FlushTask.java
   - watches creation of outFile instead of deletion
   - added requestFile and errorFile
   - no sleeping
   - a new argument in constructor; flushDir
   - a new 'start(path)' method to create requestFiles with path if path is not in StorageInfo

 + AbstractEnditNearlineStorage.java
   - added flushDir
   - cleaning the paths at start up
   - 'activate' is replaced with 'activateWithPath' for flush

 + WatchingEnditNearlineStorage.java
   - watches CREATE events in inDir and outDir (requestDir is dropped from list)
   - 2 new methods to be called in OVERFLOW event;
     polls the tasks for the files in inDir and ourDir in the order of timestamps

+++ src/main/java/org/ndgf/endit/FlushTask.java
@@ -35,6 +35,15 @@ import org.slf4j.LoggerFactory;
 
 import org.dcache.pool.nearline.spi.FlushRequest;
 
+import java.nio.file.Paths;
+import com.google.common.base.Charsets;
+
+import diskCacheV111.vehicles.StorageInfos;
+import org.dcache.vehicles.FileAttributes;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import static java.util.Arrays.asList;
 
 class FlushTask implements PollingTask<Set<URI>>
@@ -45,37 +54,72 @@ class FlushTask implements PollingTask<Set<URI>>
     private final String type;
     private final String name;
 
-    private final static Logger LOGGER = LoggerFactory.getLogger(FlushTask.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(FlushTask.class);
 
-    public FlushTask(FlushRequest request, Path outDir, String type, String name)
+    private final Path errorFile;
+    private final Path requestFile;
+    private final FileAttributes fileAttributes;
+ 
+    public FlushTask(FlushRequest request, Path outDir, Path reqDir, String type, String name)
     {
         this.type = type;
         this.name = name;
-        file = request.getFile();
+        file = request.getFile();  // deprecated
+        //file = Paths.get(request.getReplicaUri());  // dCache v3.0
         outFile = outDir.resolve(file.getName());
         pnfsId = request.getFileAttributes().getPnfsId();
+        fileAttributes = request.getFileAttributes();
+        errorFile = outDir.resolve(file.getName() + ".err");
+        requestFile = reqDir.resolve(file.getName());
     }
 
     public List<Path> getFilesToWatch()
     {
-        return asList(outFile);
+        return asList(outFile, errorFile);
     }
 
     @Override
     public Set<URI> start() throws IOException
     {
-        try {
-            Files.createLink(outFile, file.toPath());
-        } catch (FileAlreadyExistsException ignored) {
+        LOGGER.warn("FlushTask.start: {}", pnfsId.toString());
+        String si = StorageInfos.extractFrom(fileAttributes).toString();
+        Files.write(requestFile, si.getBytes(Charsets.UTF_8));
+        return null;
+    }
+
+    // new to be used with activeWithPath() to get pnfspath
+    @Override
+    public Set<URI> start(String path) throws IOException
+    {
+        LOGGER.warn("FlushTask.start: {}", pnfsId.toString());
+        String si = StorageInfos.extractFrom(fileAttributes).toString();
+        if ( ! si.contains("path=") ) {
+            if (path != null && !path.isEmpty()) {
+                StringBuffer sb = new StringBuffer("path="+path+";");
+                sb.append(si);
+                si = sb.toString();
+            }
         }
+        Files.write(requestFile, si.getBytes(Charsets.UTF_8));
         return null;
     }
 
     @Override
-    public Set<URI> poll() throws URISyntaxException
+    public Set<URI> poll() throws URISyntaxException, IOException, InterruptedException, EnditException
     {
-        if (!Files.exists(outFile)) {
-           LOGGER.debug("File " + name + " deleted");
+        if (Files.exists(errorFile)) {
+            List<String> lines;
+            try {
+                lines = Files.readAllLines(errorFile, Charsets.UTF_8);
+            } finally {
+                Files.deleteIfExists(outFile);
+                Files.deleteIfExists(errorFile);
+                Files.deleteIfExists(requestFile);
+            }
+            LOGGER.info("FlushTask.poll: request failed: {} {}", pnfsId.toString(), lines);
+            throw EnditException.create(lines);
+        }
+        if (Files.exists(outFile)) {
            URI uri = new URI(type, name, null, "bfid=" + pnfsId.toString(), null);
            // URI format: hsmType://hsmInstance/?store=storename&group=groupname&bfid=bfid  
            // <hsmType>: The type of the Tertiary Storage System  
@@ -83,6 +127,9 @@ class FlushTask implements PollingTask<Set<URI>>
            // <storename> and <groupname> : The store and group name of the file as provided by the arguments to this executable.  
            // <bfid>: The unique identifier needed to restore or remove the file if necessary.   
            LOGGER.debug("Send back uri: " + uri.toString());
+           LOGGER.warn("FlushTask.poll: file on-tape: {}", pnfsId.toString());
+           Files.deleteIfExists(outFile);
+           Files.deleteIfExists(requestFile);
            return Collections.singleton(uri);
         }
         return null;
@@ -91,6 +138,13 @@ class FlushTask implements PollingTask<Set<URI>>
     @Override
     public boolean abort() throws IOException
     {
-        return Files.deleteIfExists(outFile);
+        //return Files.deleteIfExists(outFile);
+        LOGGER.info("FlushTask.abort: {}", pnfsId.toString());  
+        if (Files.deleteIfExists(requestFile)) {
+            Files.deleteIfExists(errorFile);
+            Files.deleteIfExists(outFile);
+            return true;
+        }
+        return false;
     }
 }
